apps/api/pyproject.toml
[project]
name = "coloriage-magique-api"
version = "1.0.0"
description = "FastAPI backend for Coloriage Magique — converts photos to printable PDF coloring pages"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pillow>=10.4",
    "pillow-heif>=0.18",
    "opencv-python-headless>=4.10",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "reportlab>=4.2",
    "python-multipart>=0.0.12",
    "apscheduler>=3.10",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
